% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sctheme.R
\name{sctheme}
\alias{sctheme}
\title{Apply the Complete sctheme Brand Style}
\usage{
sctheme()
}
\value{
A list of ggplot2 theme and scale components.
}
\description{
Combines the custom plot theme with the sctheme
color and fill scales for consistent styling.
}
