% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_shares.R
\name{calculate_shares}
\alias{calculate_shares}
\title{Calculate Category Counts and Percentage Shares}
\usage{
calculate_shares(data, ...)
}
\arguments{
\item{data}{A data frame or tibble}

\item{...}{Unquoted column name(s) to group and aggregate by}
}
\value{
A summarized tibble with columns \code{n} (counts) and \code{percent} (percentage of total)
}
\description{
Takes a data frame and grouping variable(s), counts the frequency of each group,
and calculates the proportion/percentage share.
}
\examples{
\dontrun{
library(dplyr)
calculate_shares(mpg, class)
}
}
