% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_custom.R
\name{theme_custom}
\alias{theme_custom}
\title{Custom ggplot2 Theme}
\usage{
theme_custom(base_size = 11, base_family = "Inter")
}
\arguments{
\item{base_size}{Base font size in points. Default is 11.}

\item{base_family}{Base font family. Default is "Inter".}
}
\value{
A ggplot2 theme object.
}
\description{
Custom ggplot2 Theme
}
